Move test project to correct folder name
