+++ .gitignore
@@ -30,3 +30,4 @@ _ReSharper*/
 
 Source/Griffin.AdoNetFakes/Help/*
 Source/Griffin.AdoNetFakes/Fakes.shfbproj_*
+.vs/
