--- Source/Griffin.AdoNetFakes/Griffin.AdNetFakes.Tests/packages.config
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="xunit" version="1.9.1" targetFramework="net45" />
-</packages>
